move GtWrapper test to NBuildTasksTests
